requirements.txt
httpx>=0.27.0
pydantic>=2.0.0
uvicorn>=0.24.0
fastapi>=0.104.0
python-dotenv>=1.0.0
structlog>=23.2.0
jinja2>=3.1.0
requests>=2.31.0

# Tool integrations
python-trello>=0.19.0
google-api-python-client>=2.0.0
python-dateutil>=2.8.0
icalendar>=5.0.0

# Enhanced processing
numpy>=1.24.0
pandas>=2.0.0
